backend/requirements.txt
# ╔═══════════════════════════════════════════════════════════════════╗
# ║  JOBPILOT — Python Dependencies                                  ║
# ╚═══════════════════════════════════════════════════════════════════╝

# ─── Core Framework ───
fastapi==0.115.0            # Web framework (async, auto-docs)
uvicorn[standard]==0.30.6   # ASGI server
pydantic==2.9.2             # Data validation
pydantic-settings==2.5.2    # Settings management from env

# ─── Database ───
motor==3.5.1                # Async MongoDB driver
pymongo==4.8.0              # MongoDB driver (motor depends on this)

# ─── AI & NLP ───
# Using kiro-cli locally — no API library needed

# ─── Web Scraping ───
playwright==1.47.0          # Browser automation (scraping + auto-apply)
beautifulsoup4==4.12.3      # HTML parsing
lxml==5.3.0                 # Fast XML/HTML parser
httpx==0.27.2               # Async HTTP client

# ─── PDF Generation ───
weasyprint==62.3             # HTML to PDF converter
Jinja2==3.1.4               # HTML template engine

# ─── Scheduling ───
apscheduler==3.10.4         # Background job scheduler

# ─── Telegram ───
python-telegram-bot==21.5   # Telegram bot API

# ─── Utilities ───
python-dotenv==1.0.1        # .env file loading
python-multipart==0.0.9     # File upload support
aiofiles==24.1.0            # Async file operations
loguru==0.7.2               # Better logging

# ─── Development ───
ruff==0.6.5                 # Linter
pytest==8.3.3               # Testing
pytest-asyncio==0.24.0      # Async test support
